# swagger-ui/src/assets/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Api Documentation",
    "description": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Inferred Url"
    }
  ],
...
  "paths": {
    "/v8/cart": {
      "get": {
        "tags": [
          "07、购物车"
        ],
        "summary": "获取购物车详情",
        "description": "购物车-详情",
        "operationId": "findUseCartUsingGET",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "数量",
            "required": false,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userId2",
            "in": "query",
            "description": "数量",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userName",
            "in": "query",
            "description": "数量",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespEntity«CartUiEntity»"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "07、购物车"
        ],
        "summary": "获取购物车详情",
        "description": "购物车-详情",
        "operationId": "modifyUsingPOST",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "数量",
            "required": false,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "userName",
            "in": "query",
            "description": "数量",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCartEntity"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespEntity«CartUiEntity»"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddCartEntity": {
        "title": "AddCartEntity",
        "required": [
          "itemId",
          "itemName"
        ],
        "type": "object",
        "properties": {
          "itemId": {
            "type": "integer",
            "description": "编号",
            "format": "int32"
          },
          "itemName": {
            "type": "string",
            "description": "结果"
          }
        }
      },
      "CartUiEntity": {
        "title": "CartUiEntity",
        "required": [
          "items",
          "num"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "description": "列表",
            "items": {
              "$ref": "#/components/schemas/CellEntity"
            }
          },
          "num": {
            "type": "integer",
            "description": "数量",
            "format": "int32"
          }
        }
      },
      "CellEntity": {
        "title": "CellEntity",
        "required": [
          "name",
          "time"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "名称"
          },
          "time": {
            "description": "时间",
            "$ref": "#/components/schemas/Timestamp"
          }
        }
      },
...
      "RespEntity«CartUiEntity»": {
        "title": "RespEntity«CartUiEntity»",
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "执行结果的状态",
            "format": "int32"
          },
          "data": {
            "description": "正常返回的数据体,若执行失败该部分会为null，若无数据为空[]",
            "$ref": "#/components/schemas/CartUiEntity"
          },
          "message": {
            "type": "string",
            "description": "附加消息,以code为准，返回信息不一定完整"
          }
        }
      },
      "Timestamp": {
        "title": "Timestamp",
        "type": "object",
        "properties": {
          "date": {
            "type": "integer",
            "format": "int32"
          },
          "day": {
            "type": "integer",
            "format": "int32"
          },
          "hours": {
            "type": "integer",
            "format": "int32"
          },
          "minutes": {
            "type": "integer",
            "format": "int32"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "nanos": {
            "type": "integer",
            "format": "int32"
          },
          "seconds": {
            "type": "integer",
            "format": "int32"
          },
          "time": {
            "type": "integer",
            "format": "int64"
          },
          "timezoneOffset": {
            "type": "integer",
            "format": "int32"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}